+++ CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+-   Fixed crash when using `stdio.color("reset")` or `stdio.style("reset")` in a CI environment or non-interactive terminal
+
 ## `0.5.1` - February 25th, 2023
 
 ### Added